+++ cmake/sources.cmake
@@ -9,10 +9,10 @@ set(LIBEDHOC_ROOT_DIR ${CMAKE_CURRENT_LIST_DIR}/..)
 
 set(LIBEDHOC_CORE_SOURCES
     ${LIBEDHOC_ROOT_DIR}/library/core/edhoc.c
-    ${LIBEDHOC_ROOT_DIR}/library/core/edhoc_message_1.c
-    ${LIBEDHOC_ROOT_DIR}/library/core/edhoc_message_2.c
-    ${LIBEDHOC_ROOT_DIR}/library/core/edhoc_message_3.c
-    ${LIBEDHOC_ROOT_DIR}/library/core/edhoc_message_4.c
+    ${LIBEDHOC_ROOT_DIR}/library/core/classic/edhoc_classic_message_1.c
+    ${LIBEDHOC_ROOT_DIR}/library/core/classic/edhoc_classic_message_2.c
+    ${LIBEDHOC_ROOT_DIR}/library/core/classic/edhoc_classic_message_3.c
+    ${LIBEDHOC_ROOT_DIR}/library/core/classic/edhoc_classic_message_4.c
     ${LIBEDHOC_ROOT_DIR}/library/core/edhoc_message_error.c
     ${LIBEDHOC_ROOT_DIR}/library/core/edhoc_exporter.c
     ${LIBEDHOC_ROOT_DIR}/library/core/edhoc_cbor_internal.c