+++ library/core/edhoc.c
@@ -32,6 +32,7 @@ LOG_MODULE_REGISTER(libedhoc, CONFIG_LIBEDHOC_LOG_LEVEL);
 /* EDHOC internal headers: */
 #include "edhoc_context_internal.h"
 #include "edhoc_key_slot_internal.h"
+#include "edhoc_classic_internal.h"
 #include "edhoc_connection_id_internal.h"
 #include "edhoc_backend_log.h"
 
@@ -371,4 +372,56 @@ int edhoc_error_get_cipher_suites(const struct edhoc_context *ctx,
 			ctx->negotiation.peer_cipher_suite.entry[i].value;
 
 	return EDHOC_SUCCESS;
+}
+
+int edhoc_message_1_compose(struct edhoc_context *ctx, uint8_t *msg_1,
+			    size_t msg_1_size, size_t *msg_1_len)
+{
+	return edhoc_classic_message_1_compose(ctx, msg_1, msg_1_size,
+					       msg_1_len);
+}
+
+int edhoc_message_1_process(struct edhoc_context *ctx, const uint8_t *msg_1,
+			    size_t msg_1_len)
+{
+	return edhoc_classic_message_1_process(ctx, msg_1, msg_1_len);
+}
+
+int edhoc_message_2_compose(struct edhoc_context *ctx, uint8_t *msg_2,
+			    size_t msg_2_size, size_t *msg_2_len)
+{
+	return edhoc_classic_message_2_compose(ctx, msg_2, msg_2_size,
+					       msg_2_len);
+}
+
+int edhoc_message_2_process(struct edhoc_context *ctx, const uint8_t *msg_2,
+			    size_t msg_2_len)
+{
+	return edhoc_classic_message_2_process(ctx, msg_2, msg_2_len);
+}
+
+int edhoc_message_3_compose(struct edhoc_context *ctx, uint8_t *msg_3,
+			    size_t msg_3_size, size_t *msg_3_len)
+{
+	return edhoc_classic_message_3_compose(ctx, msg_3, msg_3_size,
+					       msg_3_len);
+}
+
+int edhoc_message_3_process(struct edhoc_context *ctx, const uint8_t *msg_3,
+			    size_t msg_3_len)
+{
+	return edhoc_classic_message_3_process(ctx, msg_3, msg_3_len);
+}
+
+int edhoc_message_4_compose(struct edhoc_context *ctx, uint8_t *msg_4,
+			    size_t msg_4_size, size_t *msg_4_len)
+{
+	return edhoc_classic_message_4_compose(ctx, msg_4, msg_4_size,
+					       msg_4_len);
+}
+
+int edhoc_message_4_process(struct edhoc_context *ctx, const uint8_t *msg_4,
+			    size_t msg_4_len)
+{
+	return edhoc_classic_message_4_process(ctx, msg_4, msg_4_len);
 }